+++ samples/Logfella.Samples.MVC/Controllers/HomeController.cs
@@ -7,7 +7,7 @@ namespace Logfella.Samples.MVC.Controllers
     [ApiController]
     public class HomeController : ControllerBase
     {
-        // Inject an ILogger<T> as normal.
+        // Inject an ILogger<T> object as normal.
         // The ILogger interface is somewhat limited when it comes
         // to logging to GCP, as Microsoft.Extensions.Logging
         // has a smaller set of severity levels and doesn't set
@@ -46,13 +46,14 @@ namespace Logfella.Samples.MVC.Controllers
             // Alternatively, if you can't be bothered to inject a logger
             // into every single class across your entire application then
             // you can use the static Log instance instead.
-            // Regardless which option you use (ILogger vs. ILogWriter vs. Log)
+            // Regardless which option you choose to use (ILogger vs. ILogWriter vs. Log)
             // you'll always be using the same global or per request instance under
             // the hood and therefore not miss out on features such as correlating
             // logs or pre-configured labels, severity, and more.
             // The static Log instance is mainly the preferred option for
-            // functional applications in F#, where constructor DI is not idiomatic.
+            // functional applications in F#, where constructor DI is not that common.
             Log.Info("Foo");
+
             return "foo";
         }
     }