+++ samples/Logfella.Samples.MVC/Program.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using Logfella.Adapters;
 using Logfella.LogWriters;
@@ -8,42 +9,46 @@ namespace Logfella.Samples.MVC
 {
     public static class Program
     {
+        // Dummy data, in a real world app you'd load these values
+        // from a config source or dynamically read from the .csproj
+        // settings:
+        private const string AppName = "SampleAppCSharp";
+        private const string AppVersion = "v1.0.0";
+
+        // Generate a new GoogleCloudLogWriter with the help of
+        // the builder methods by dynamically adding more
+        // features to the instance. These builder methods allow
+        // one to add features or configurations based on conditions
+        // in a nice fluent way:
+        public static readonly GoogleCloudLogWriter DefaultLogWriter =
+            GoogleCloudLogWriter
+                .Create(Severity.Debug) // Min severity to write logs
+                .AddServiceContext(     // Configure service context for Google Error Reporting
+                    AppName,
+                    AppVersion)
+                .AddLabels(
+                    new Dictionary<string, string>
+                    {                                // Configure additional labels which will be
+                        {"appName", AppName},        // automatically added to all log entries.
+                        {"appVersion", AppVersion}   // Use labels to make log filtering & analysis easier
+                    });                              // (e.g. which with app version did new errors start?)
+
         public static void Main(string[] args)
         {
-            // Configure an ILogWriter instance for this application.
-            // This method should only be called once during application startup.
-            // Normally a program would decide which kind of logger it wants to
-            // create based on environment settings during the start-up procedures.
-            //
-            // In an ASP.NET Core application you can configure additional "per-request"
-            // ILogWriter instances by configuring additional middleware inside Startup.cs,
-            // however those request-based log writers only exist for the scope of a HTTP
-            // request and any logs which need to be written outside the ASP.NET Core request
-            // pipeline will continue using the global log writer configured here:
-            Log.SetGlobalLogWriter(
-                new GoogleCloudLogWriter(
-                    Severity.Debug,                            // Set the minimum log level
-                    "SampleAppForCSharp",          // The service name and service version are being
-                    "v1.0.0",                     // used for Google Error Reporting.
-                    false,              // Set to true if Google Cloud Logging should determine the log timestamp itself.
-                    new Dictionary<string, string>
-                    {
-                        { "appName", "SampleAppForCSharp" },   // Labels which all logs will be decorated with.
-                        { "appVersion", "v1.0.0" },            // Use this as a mechanism to easily filter logs in GCP.
-                        { "loggerType", "global" },            // Example, set the version of the service to easily determine
-                        { "foo", "bar" }                       // which deployment might have caused an influx of errors, etc.
-                    }));
+            // Set the default ILogWriter instance.
+            // This method should ideally only be invoked once during application startup:
+            Log.SetDefaultLogWriter(DefaultLogWriter);
+
+            // Write a log entry:
+            Log.Info("Starting web server...");
 
             StartWebServer(args);
         }
 
         private static void StartWebServer(string[] args) =>
             Host.CreateDefaultBuilder(args)
-                // Configure Microsoft.Extensions.Logging to use Logfella
-                // when creating new ILogger or ILogger<T> instances.
-                //
-                // This is not necessarily required, but then logs from other middleware
-                // will not write to the same log writer as the one set above.
+                // This is only required to configure an adapter so that `Microsoft.Extensions.Logging.ILogger`
+                // objects will use the above configured ILogWriter for all logs:
                 .UseLogfella()
                 .ConfigureWebHost(
                     webBuilder =>