+++ samples/Logfella.Samples.MVC/Startup.cs
@@ -1,7 +1,5 @@
 using System;
-using System.Collections.Generic;
 using Logfella.AspNetCore;
-using Logfella.LogWriters;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
@@ -18,45 +16,33 @@ namespace Logfella.Samples.MVC
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            // Configure a factory method to generate a new ILogWriter instance
-            // for each individual incoming HTTP request of the ASP.NET Core pipeline.
-            // This middleware is entirely optional and not required to configure
-            // Logfella in general. It mainly serves the purpose of being able to
-            // set a unique "correlation ID" on an ILogWriter in order to allow
-            // request based log aggregation in the Google Cloud Logging console
-            // (or any other log aggregation tool in that matter):
-            app.UsePerRequestLogWriter(
+            // This is an optional middleware which can be configured to create a new
+            // `ILogWriter` instance for each incoming HTTP request. This makes it possible
+            // to decorate all log entries with the HTTP context (see `AddHttpContext(ctx)`)
+            // and to set a correlation ID for all log entries which belong to the same
+            // HTTP request (see `AddCorrelationId(..., ...)`).
+            //
+            // Because this middleware requires a factory method (Func<HttpContext,ILogWriter>)
+            // one has full control over how or which `ILogWriter` to use and to inspect an incoming
+            // HttpRequest in order to dynamically configure the instance (e.g. check the HTTP
+            // headers for an X-Correlation-Id header and use that value as `requestId`):
+            app.UseLogWriterPerRequest(
                 ctx =>
-                    new GoogleCloudLogWriter(
-                        Severity.Debug,
-                        "SampleAppForCSharp",
-                        "v1.0.0",
-                        false,
-                        new Dictionary<string, string>
-                            {
-                                { "appName", "SampleAppForCSharp" },
-                                { "appVersion", "v1.0.0" },
-                                { "foo", "bar" }
-                            },
-
-                        // The key/property under which the correlation ID
-                        // will be serialised in the structured logs.
-                        // The default is 'correlationId' but it can be useful
-                        // to set it to a different value if that name is already
-                        // taken by another property which might occur in log entries:
-                        "correlationId",
-
-                        // Configure a correlation ID for all logs as part of
-                        // a single HTTP request. Either generate a new unique ID
-                        // (e.g. Guid.NewGuid.ToString()) or inspect the HTTP request
-                        // (e.g. X-Correlation-Id HTTP header) to retrieve an existing
-                        // correlation ID from another service in the chain of commands:
-                        ctx.Request.Query.ContainsKey("cid")
-                            ? ctx.Request.Query["cid"].ToString()
-                            : Guid.NewGuid().ToString("N"))
-
-                        // Include additional HTTP request data with each log entry (optional):
-                        .IncludeHttpRequest(ctx));
+                    // Note that one can use an already existing and pre-configured GoogleCloudLogWriter
+                    // to extend it with additional features without having to re-create the entire
+                    // object again. The GoogleCloudLogWriter is immutable and the builder methods
+                    // always create a new instance, which makes it safe to use it in this way:
+                    Program.DefaultLogWriter
+                        .AddLabel("loggerType", "requestLogger")
+                        .AddHttpContext(ctx)
+                        .AddCorrelationId(
+                            // Correlation ID/Request ID can either be inherited from an incoming
+                            // web request in a distributed system or a new one generated:
+                            ctx.Request.Headers.ContainsKey("X-Correlation-Id")
+                                ? ctx.Request.Headers["X-Correlation-Id"].ToString()
+                                : Guid.NewGuid().ToString("N"),
+                            "requestId")); // <-- Change the key name if the default `correlationId`
+                                                //     is already used for something else.
 
             if (env.IsDevelopment())
                 app.UseDeveloperExceptionPage();