+++ src/Logfella/Adapters/AdapterExtensions.cs
@@ -1,10 +1,16 @@
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 
 namespace Logfella.Adapters
 {
-    public static class HostBuilderExtensions
+    public static class AdapterExtensions
     {
+        public static ILoggingBuilder UseLogfella(this ILoggingBuilder builder) =>
+            builder
+                .ClearProviders()
+                .AddProvider(new LoggerAdapterProvider());
+
         public static IHostBuilder UseLogfella(this IHostBuilder builder) =>
             builder
                 .ConfigureServices(