+++ src/Logfella/LogWriter.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using Logfella.Extensions;
 
 namespace Logfella
 {
@@ -12,8 +11,15 @@ namespace Logfella
         /// </summary>
         protected readonly Severity MinSeverity;
 
-        private readonly string _correlationIdKey;
-        private readonly string _correlationId;
+        /// <summary>
+        /// Property name under which a possible correlation ID will be logged.
+        /// </summary>
+        protected readonly string CorrelationIdKey;
+
+        /// <summary>
+        /// Correlation ID value to aggregate a set of related logs (e.g. per HTTP request).
+        /// </summary>
+        protected readonly string CorrelationId;
 
         /// <summary>
         /// Initialises a new instance of a LogWriter with a minimum severity level.
@@ -24,8 +30,24 @@ namespace Logfella
             string correlationId = "")
         {
             MinSeverity = minSeverity;
-            _correlationIdKey = correlationIdKey;
-            _correlationId = correlationId;
+            CorrelationIdKey = correlationIdKey;
+            CorrelationId = correlationId;
+        }
+
+        /// <summary>
+        /// <para>Convenience method to type cast into `ILogWriter` interface.</para>
+        /// <para>This method is typically redundant in C# and mostly useful for F# applications only.</para>
+        /// </summary>
+        public ILogWriter AsLogWriter() => this;
+
+        private static IDictionary<TKey, TValue> ToDict<TKey, TValue>(IEnumerable<Tuple<TKey, TValue>> array)
+        {
+            var dict = new Dictionary<TKey, TValue>();
+
+            foreach (var (key, value) in array)
+                dict.Add(key, value);
+
+            return dict;
         }
 
         /// <summary>
@@ -49,10 +71,10 @@ namespace Logfella
             if (severity < MinSeverity)
                 return;
 
-            if (!string.IsNullOrEmpty(_correlationId))
+            if (!string.IsNullOrEmpty(CorrelationId))
             {
                 data ??= new Dictionary<string, object>();
-                data.Add(_correlationIdKey, _correlationId);
+                data.Add(CorrelationIdKey, CorrelationId);
             }
 
             WriteLog(severity, message, data, ex);
@@ -64,11 +86,11 @@ namespace Logfella
 
         /// <inheritdoc />
         public void Default(string message, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Default, message, data.ToDict());
+            WriteLogIfSeverity(Severity.Default, message, ToDict(data));
 
         /// <inheritdoc />
         public void Default(string message, Exception ex, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Default, message, data.ToDict(), ex);
+            WriteLogIfSeverity(Severity.Default, message, ToDict(data), ex);
 
         /// <inheritdoc />
         public void Default(string message, IDictionary<string, object> data, Exception ex = null) =>
@@ -76,11 +98,11 @@ namespace Logfella
 
         /// <inheritdoc />
         public void Debug(string message, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Debug, message, data.ToDict());
+            WriteLogIfSeverity(Severity.Debug, message, ToDict(data));
 
         /// <inheritdoc />
         public void Debug(string message, Exception ex, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Debug, message, data.ToDict(), ex);
+            WriteLogIfSeverity(Severity.Debug, message, ToDict(data), ex);
 
         /// <inheritdoc />
         public void Debug(string message, IDictionary<string, object> data, Exception ex = null) =>
@@ -88,11 +110,11 @@ namespace Logfella
 
         /// <inheritdoc />
         public void Info(string message, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Info, message, data.ToDict());
+            WriteLogIfSeverity(Severity.Info, message, ToDict(data));
 
         /// <inheritdoc />
         public void Info(string message, Exception ex, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Info, message, data.ToDict(), ex);
+            WriteLogIfSeverity(Severity.Info, message, ToDict(data), ex);
 
         /// <inheritdoc />
         public void Info(string message, IDictionary<string, object> data, Exception ex = null) =>
@@ -100,11 +122,11 @@ namespace Logfella
 
         /// <inheritdoc />
         public void Notice(string message, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Notice, message, data.ToDict());
+            WriteLogIfSeverity(Severity.Notice, message, ToDict(data));
 
         /// <inheritdoc />
         public void Notice(string message, Exception ex, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Notice, message, data.ToDict(), ex);
+            WriteLogIfSeverity(Severity.Notice, message, ToDict(data), ex);
 
         /// <inheritdoc />
         public void Notice(string message, IDictionary<string, object> data, Exception ex = null) =>
@@ -112,11 +134,11 @@ namespace Logfella
 
         /// <inheritdoc />
         public void Warning(string message, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Warning, message, data.ToDict());
+            WriteLogIfSeverity(Severity.Warning, message, ToDict(data));
 
         /// <inheritdoc />
         public void Warning(string message, Exception ex, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Warning, message, data.ToDict(), ex);
+            WriteLogIfSeverity(Severity.Warning, message, ToDict(data), ex);
 
         /// <inheritdoc />
         public void Warning(string message, IDictionary<string, object> data, Exception ex = null) =>
@@ -124,11 +146,11 @@ namespace Logfella
 
         /// <inheritdoc />
         public void Error(string message, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Error, message, data.ToDict());
+            WriteLogIfSeverity(Severity.Error, message, ToDict(data));
 
         /// <inheritdoc />
         public void Error(string message, Exception ex, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Error, message, data.ToDict(), ex);
+            WriteLogIfSeverity(Severity.Error, message, ToDict(data), ex);
 
         /// <inheritdoc />
         public void Error(string message, IDictionary<string, object> data, Exception ex = null) =>
@@ -136,11 +158,11 @@ namespace Logfella
 
         /// <inheritdoc />
         public void Critical(string message, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Critical, message, data.ToDict());
+            WriteLogIfSeverity(Severity.Critical, message, ToDict(data));
 
         /// <inheritdoc />
         public void Critical(string message, Exception ex, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Critical, message, data.ToDict(), ex);
+            WriteLogIfSeverity(Severity.Critical, message, ToDict(data), ex);
 
         /// <inheritdoc />
         public void Critical(string message, IDictionary<string, object> data, Exception ex = null) =>
@@ -148,11 +170,11 @@ namespace Logfella
 
         /// <inheritdoc />
         public void Alert(string message, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Alert, message, data.ToDict());
+            WriteLogIfSeverity(Severity.Alert, message, ToDict(data));
 
         /// <inheritdoc />
         public void Alert(string message, Exception ex, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Alert, message, data.ToDict(), ex);
+            WriteLogIfSeverity(Severity.Alert, message, ToDict(data), ex);
 
         /// <inheritdoc />
         public void Alert(string message, IDictionary<string, object> data, Exception ex = null) =>
@@ -160,11 +182,11 @@ namespace Logfella
 
         /// <inheritdoc />
         public void Emergency(string message, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Emergency, message, data.ToDict());
+            WriteLogIfSeverity(Severity.Emergency, message, ToDict(data));
 
         /// <inheritdoc />
         public void Emergency(string message, Exception ex, params Tuple<string, object>[] data) =>
-            WriteLogIfSeverity(Severity.Emergency, message, data.ToDict(), ex);
+            WriteLogIfSeverity(Severity.Emergency, message, ToDict(data), ex);
 
         /// <inheritdoc />
         public void Emergency(string message, IDictionary<string, object> data, Exception ex = null) =>