+++ src/Logfella/LogWriters/GoogleCloudLogWriter.cs
@@ -1,22 +1,30 @@
 using System;
 using System.Collections.Generic;
-using System.Text;
+using System.Linq;
 using System.Text.Json;
-using Logfella.Extensions;
 using Microsoft.AspNetCore.Http;
 
 namespace Logfella.LogWriters
 {
     public sealed class GoogleCloudLogWriter : LogWriter
     {
-        public class HttpRequest
+        // ------------------------------------------
+        // Sub types to serialise extra log data
+        // ------------------------------------------
+
+        public class HttpRequestContext
         {
-            public HttpRequest(HttpContext ctx)
+            public HttpRequestContext(HttpContext ctx)
             {
                 var req = ctx.Request;
+                var path = req.PathBase.HasValue || req.Path.HasValue
+                    ? req.PathBase.Value + req.Path.Value
+                    : "/";
+                var requestUrl =
+                    $"{req.Scheme}://{req.Host.Value}{path}{req.QueryString.Value}";
 
                 RequestMethod = req.Method.ToUpper();
-                RequestUrl = ctx.RequestUrl();
+                RequestUrl = requestUrl;
                 RequestSize =
                     req.ContentLength.HasValue
                         ? req.ContentLength.Value.ToString()
@@ -33,6 +41,37 @@ namespace Logfella.LogWriters
                         : "";
                 Protocol = req.Protocol;
             }
+
+            private HttpRequestContext(
+                string requestMethod,
+                string requestUrl,
+                string requestSize,
+                string userAgent,
+                string remoteIp,
+                string serverIp,
+                string referer,
+                string protocol)
+            {
+                RequestMethod = requestMethod;
+                RequestUrl = requestUrl;
+                RequestSize = requestSize;
+                UserAgent = userAgent;
+                RemoteIp = remoteIp;
+                ServerIp = serverIp;
+                Referer = referer;
+                Protocol = protocol;
+            }
+
+            public HttpRequestContext DeepClone() =>
+                new HttpRequestContext(
+                    RequestMethod,
+                    RequestUrl,
+                    RequestSize,
+                    UserAgent,
+                    RemoteIp,
+                    ServerIp,
+                    Referer,
+                    Protocol);
 
             public string RequestMethod { get; }
             public string RequestUrl { get; }
@@ -63,6 +102,10 @@ namespace Logfella.LogWriters
             public ErrorContext InnerError { get; }
         }
 
+        // ------------------------------------------
+        // JSON Serializer settings
+        // ------------------------------------------
+
         private readonly JsonSerializerOptions _jsonOptions =
             new JsonSerializerOptions
             {
@@ -74,11 +117,65 @@ namespace Logfella.LogWriters
                 PropertyNamingPolicy = JsonNamingPolicy.CamelCase
             };
 
+        // ------------------------------------------
+        // Google Cloud LogEntry Keys & Values
+        // ------------------------------------------
+
+        private static class Keys
+        {
+            // Property names which the GCP log analyzer will
+            // automatically pick up from the jsonPayload:
+            // (See: https://cloud.google.com/logging/docs/agent/configuration#special-fields)
+
+            public const string Severity = "severity";
+            public const string Time = "time";
+            public const string ServiceName = "serviceContext.service";
+            public const string ServiceVersion = "serviceContext.version";
+            public const string Message = "message";
+            public const string Type = "@type";
+            public const string HttpRequest = "httpRequest";
+            public const string Labels = "logging.googleapis.com/labels";
+        }
+
+        private static class Values
+        {
+            public const string ErrorType =
+                "type.googleapis.com/google.devtools.clouderrorreporting.v1beta1.ReportedErrorEvent";
+        }
+
+        // ------------------------------------------
+        // Private fields
+        // ------------------------------------------
+
+        private const string DefaultCorrelationIdKey = "correlationId";
         private readonly string _serviceName;
         private readonly string _serviceVersion;
         private readonly bool _useGoogleCloudTimestamp;
         private readonly Dictionary<string, string> _labels;
-        private HttpRequest _httpRequest;
+        private readonly HttpRequestContext _httpRequestContext;
+
+        // ------------------------------------------
+        // Private static helper methods for immutability
+        // ------------------------------------------
+
+        private static Dictionary<string, string> Clone(IDictionary<string, string> labels) =>
+            labels.Keys.ToDictionary(
+                key => key,
+                key => labels[key]);
+
+        private static Dictionary<string, string> CloneAndAdd(
+            IDictionary<string, string> labels,
+            params (string key, string value)[] additionalLabels)
+        {
+            var clonedLabels = Clone(labels);
+            foreach (var (key, value) in additionalLabels)
+                clonedLabels.Add(key, value);
+            return clonedLabels;
+        }
+
+        // ------------------------------------------
+        // Main public constructor
+        // ------------------------------------------
 
         /// <summary>
         /// Initialises a new GoogleCloudLogWriter object.
@@ -90,14 +187,16 @@ namespace Logfella.LogWriters
         /// <param name="labels">Optional labels to attach to all log entries.</param>
         /// <param name="correlationIdKey">The key under which a correlationId will be logged.</param>
         /// <param name="correlationId">An ID to correlate a set of logs (e.g. per HTTP request).</param>
+        /// <param name="ctx">Additional HTTP data to be set for a per-request created logger.</param>
         public GoogleCloudLogWriter(
             Severity minSeverity,
             string serviceName = null,
             string serviceVersion = null,
             bool useGoogleCloudTimestamp = false,
             IDictionary<string, string> labels = null,
-            string correlationIdKey = "correlationId",
-            string correlationId = "")
+            string correlationIdKey = DefaultCorrelationIdKey,
+            string correlationId = null,
+            HttpContext ctx = null)
             : base(
                 minSeverity,
                 correlationIdKey,
@@ -108,27 +207,147 @@ namespace Logfella.LogWriters
             _useGoogleCloudTimestamp = useGoogleCloudTimestamp;
 
             if (labels != null)
-                _labels = new Dictionary<string, string>(labels);
+                _labels = Clone(labels);
+
+            if (ctx != null)
+                _httpRequestContext = new HttpRequestContext(ctx);
         }
+
+        // ------------------------------------------
+        // Private constructor and public Clone method
+        // ------------------------------------------
+
+        private GoogleCloudLogWriter(
+            Severity minSeverity,
+            string serviceName,
+            string serviceVersion,
+            bool useGoogleCloudTimestamp,
+            IDictionary<string, string> labels,
+            string correlationIdKey,
+            string correlationId,
+            HttpRequestContext requestCtx)
+            : this(
+                minSeverity,
+                serviceName,
+                serviceVersion,
+                useGoogleCloudTimestamp,
+                labels,
+                correlationIdKey,
+                correlationId)
+        {
+            _httpRequestContext = requestCtx.DeepClone();
+        }
+
+        public GoogleCloudLogWriter DeepClone() =>
+            new GoogleCloudLogWriter(
+                MinSeverity,
+                _serviceName,
+                _serviceVersion,
+                _useGoogleCloudTimestamp,
+                // Labels will be deep cloned in the ctor:
+                _labels,
+                CorrelationIdKey,
+                CorrelationId,
+                // The requestContext will be cloned in the ctor:
+                _httpRequestContext);
+
+        // ------------------------------------------
+        // Private constructor and public builder methods
+        // ------------------------------------------
+
+        public static GoogleCloudLogWriter Create(Severity minSeverity) =>
+            new GoogleCloudLogWriter(minSeverity);
+
+        public GoogleCloudLogWriter AddServiceContext(string serviceName, string serviceVersion) =>
+            new GoogleCloudLogWriter(
+                MinSeverity,
+                serviceName,
+                serviceVersion,
+                _useGoogleCloudTimestamp,
+                _labels,
+                CorrelationIdKey,
+                CorrelationId,
+                _httpRequestContext);
+
+        public GoogleCloudLogWriter UseGoogleCloudTimestamp() =>
+            new GoogleCloudLogWriter(
+                MinSeverity,
+                _serviceName,
+                _serviceVersion,
+                true,
+                _labels,
+                CorrelationIdKey,
+                CorrelationId,
+                _httpRequestContext);
+
+        public GoogleCloudLogWriter AddLabels(IDictionary<string, string> labels) =>
+            new GoogleCloudLogWriter(
+                MinSeverity,
+                _serviceName,
+                _serviceVersion,
+                _useGoogleCloudTimestamp,
+                labels,
+                CorrelationIdKey,
+                CorrelationId,
+                _httpRequestContext);
+
+        public GoogleCloudLogWriter AddLabel(string key, string value) =>
+            new GoogleCloudLogWriter(
+                MinSeverity,
+                _serviceName,
+                _serviceVersion,
+                _useGoogleCloudTimestamp,
+                CloneAndAdd(_labels, (key, value)),
+                CorrelationIdKey,
+                CorrelationId,
+                _httpRequestContext);
+
+        public GoogleCloudLogWriter AddCorrelationId(
+            string correlationId,
+            string key = DefaultCorrelationIdKey) =>
+            new GoogleCloudLogWriter(
+                MinSeverity,
+                _serviceName,
+                _serviceVersion,
+                _useGoogleCloudTimestamp,
+                _labels,
+                key,
+                correlationId,
+                _httpRequestContext);
+
+        public GoogleCloudLogWriter AddHttpContext(HttpContext httpContext) =>
+            new GoogleCloudLogWriter(
+                MinSeverity,
+                _serviceName,
+                _serviceVersion,
+                _useGoogleCloudTimestamp,
+                _labels,
+                CorrelationIdKey,
+                CorrelationId,
+                httpContext);
 
         /// <summary>
-        /// If true then the JSON log output will be written in indented format for better human readability.
-        /// <para>Use this setting to debug log output during development.</para>
+        /// Use this setting to debug log output during development.
         /// </summary>
-        public bool PrettifyOutput
+        public GoogleCloudLogWriter PrettifyOutputForDebugging()
         {
-            set => _jsonOptions.WriteIndented = value;
+            var instance =
+                new GoogleCloudLogWriter(
+                    MinSeverity,
+                    _serviceName,
+                    _serviceVersion,
+                    _useGoogleCloudTimestamp,
+                    _labels,
+                    CorrelationIdKey,
+                    CorrelationId,
+                    _httpRequestContext);
+            instance._jsonOptions.WriteIndented = true;
+            return instance;
         }
 
-        public ILogWriter IncludeHttpRequest(HttpContext ctx)
-        {
-            if (ctx == null)
-                throw new ArgumentNullException(nameof(ctx));
-
-            _httpRequest = new HttpRequest(ctx);
-
-            return this;
-        }
+        // ------------------------------------------
+        // Main logging method
+        // ------------------------------------------
 
         // In the GCP any logs written directly to stdout will be picked up by
         // Google StackDriver Logging and indexed correctly.
@@ -143,35 +362,36 @@ namespace Logfella.LogWriters
             IDictionary<string, object> data,
             Exception ex)
         {
-            var msg = new StringBuilder();
-            msg.Append(message);
-
             var logEntry = new Dictionary<string, object>
-            {
-                {"message", msg.ToString()},
-                {"severity", severity.String()}
-            };
+                {
+                    {Keys.Severity, severity.String()}
+                };
 
             if (!_useGoogleCloudTimestamp)
-                logEntry.Add("time", DateTime.UtcNow.ToString("o"));
+                logEntry.Add(Keys.Time, DateTime.UtcNow.ToString("o"));
 
             if (string.IsNullOrEmpty(_serviceName))
-                logEntry.Add("serviceContext.service", _serviceName);
+                logEntry.Add(Keys.ServiceName, _serviceName);
 
             if (string.IsNullOrEmpty(_serviceVersion))
-                logEntry.Add("serviceContext.version", _serviceVersion);
+                logEntry.Add(Keys.ServiceVersion, _serviceVersion);
 
             if (ex != null)
             {
-                logEntry.Add("@type", "type.googleapis.com/google.devtools.clouderrorreporting.v1beta1.ReportedErrorEvent");
+                logEntry.Add(Keys.Message, $"{message}\n\nException:\n\n{ex}");
+                //logEntry.Add(Keys.Type, Values.ErrorType);
                 logEntry.Add("error", new ErrorContext(ex));
             }
+            else
+            {
+                logEntry.Add(Keys.Message, message);
+            }
 
-            if (_httpRequest != null)
-                logEntry.Add("httpRequest", _httpRequest);
+            if (_httpRequestContext != null)
+                logEntry.Add(Keys.HttpRequest, _httpRequestContext);
 
             if (_labels != null && _labels.Count > 0)
-                logEntry.Add("logging.googleapis.com/labels", _labels);
+                logEntry.Add(Keys.Labels, _labels);
 
             if (data != null && data.Keys.Count > 0)
                 logEntry.Add("data", data as Dictionary<string, object>);